visual studio 6.0 compatibility
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@917 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ include/irrMath.h
@@ -234,7 +234,7 @@ namespace core
 
 	inline f32 round_( f32 x )
 	{
-		return floor( x + 0.5f );
+		return (f32)floor( x + 0.5f );
 	}
 
 	REALINLINE void clearFPUException ()
